% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModificationResults.R
\name{plotTopKmers}
\alias{plotTopKmers}
\title{Plot frequency bar plot of top n kmers}
\usage{
plotTopKmers(modResults, numKmers = 10)
}
\arguments{
\item{modResults}{RNA modification detection results in a dataframe}

\item{numKmers}{Integer value specifying how many top kmers to display}
}
\value{
a bar plot of the frequencies of the top n kmers
}
\description{
Plot frequency bar plot of top n kmers
}
\examples{

plotTopKmers(RnaModificationResults, numKmers = 10)

}
\references{
\enumerate{
\item Pratanwanich, P. N., Yao, F., Chen, Y., Koh, C. W. Q., Wan, Y. K., Hendra, C., Poon, P., Goh,
Y. T., Yap, P. M. L., Chooi, J. Y., Chng, W. J., Ng, S. B., Thiery, A., Goh, W. S. S., & Göke, J. (2021).
Identification of differential RNA modifications from nanopore direct RNA sequencing with xPore.
Nature Biotechnology,39(11), 1394–1402. https://doi.org/10.1038/s41587-021-00949-w
}
}
\author{
{Dae-won Gong, \email{daewon.gong@mail.utoronto.ca}}
}
