% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNanoPlotR.R
\name{runNanoPlotR}
\alias{runNanoPlotR}
\title{Launch Shiny App For Package NanoPlotR}
\usage{
runNanoPlotR()
}
\value{
No return value but opens up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The shiny app allows user to input a RNA modification result file,
then app will plot the visualization plots.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
NanoPlotR::runNanoPlotR()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
\author{
Dae-won Gong \email{daewon.gong@mail.utoronto.ca}
}
