% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModificationResults.R
\name{plotModHist}
\alias{plotModHist}
\title{Plot histogram of modification rates.}
\usage{
plotModHist(modResults)
}
\arguments{
\item{modResults}{RNA modification detection results in a data frame.}
}
\value{
Histograms of all modification rates in 1 panel.
}
\description{
Plot histogram of modification rates.
}
\examples{

plotModHist(RnaModificationResults)

}
\references{
\enumerate{
\item Frank E Harrell Jr (2021). Hsmisc: Harrell Miscellaneous.
R package version 4.6.0. https://CRAN.R-project.org/package=Hmisc
\item Pratanwanich, P. N., Yao, F., Chen, Y., Koh, C. W. Q., Wan, Y. K., Hendra, C., Poon, P., Goh,
Y. T., Yap, P. M. L., Chooi, J. Y., Chng, W. J., Ng, S. B., Thiery, A., Goh, W. S. S., & Göke, J. (2021).
Identification of differential RNA modifications from nanopore direct RNA sequencing with xPore.
Nature Biotechnology,39(11), 1394–1402. https://doi.org/10.1038/s41587-021-00949-w
}
}
\author{
{Dae-won Gong, \email{daewon.gong@mail.utoronto.ca}}
}
